Add `get_cairo_version` xtask
